Library (data.h): two functions to allocate and free gal_data_t **
In many contexts, it is necessary to pre-allocate a list of pointers to
datasets which are then filled at later stages of the processing. With the
two new 'gal_data_array_ptr_calloc' and 'gal_data_array_ptr_free'
functions, it is now very easy to allocate such an array and later free it.