Library (cosmology.h) and CosmicCalulator: velocity to z conversion
Until now the only inputs to CosmicCalculator were the redshift and the
observed wavelength. But in some scenarios, we have the velocity of the
object instead and although the conversion isn't too complicated, it is
annoying!

With this commit, it is now possible to give the input redshift of
CosmicCalculator through the '--velocity' option (value will be
interpretted to be in units of km/s) and the used velocity in that run can
be printed out with the new '--usedvelocity' option. Also, in the general
information printed, we now also print the velocity.