Library (fits.h): can read tables with TNULL before TFORM
Until now, when a FITS table had the 'TNULL' (value for blank rows) keyword
before the 'TFORM' keyword (value containing the numerical data type), that
table couldn't be read by any of the programs. The problem was that we
needed to know the type of the column to allocate space for the value of
'TNULL'.

With this commit, when this happens, the keyname and value are stored in a
list and inserted into the internal data structure after all other keywords
have been read.

This bug was reported by Samane Raji.

This fixes bug #59019.