Library (kdtree.h): new k-d tree space-partitioning tree library
With this commit, a new k-d tree library is added in Gnuasto, that is
written from scratch for optimal memory and CPU usage. It currently
contains two higher level functions for tree creation and nearest neighbour
searches.

This is particularly useful for matching multidimensional points, which
will be used in the future when the quad matching algorithm will be
integrated for astrometry in the 'match' program. It can also be used for
more optimal and multi-threaded matching of large catalogs.