ConvertType: new sls-inverse colormap, good for printing
The SLS colormap is a good color map for viewing on the monitor, but for
printing and some types of visualizations, its very dark background
decreases its usefulness.

With this commit, there is now an inverse SLS color map (called
'sls-inverse') to keep the advantages of SLS, but good for scenarios where
a white background can help.