Building: added check for systems without OBSFIX in their WCSLIB
Until now, we were simply assuming that the host WCSLIB has the 'OBSFIX'
keyword in the newly added step to run 'wcsfix' on the extracted WCS. But
it is a relatively recent keyword in WCSLIB that isn't available on some
systems, causing a crash.

With this commit, a configure-time check has been added and if this macro
isn't present, then the part that uses it won't be compiled.