MakeProfiles: only print status of profiles when they are <50
Until now, MakeProfiles would print a line for every built profile, and say
that there are "N" profiles left. However, in many scenarios (in paritcular
aperture profiles that are small!) the printing itself can be slower than
the creation time of the profile and this will un-necessarily slow-down the
program.

With this commit, MakeProfiles now only prints the profile status if there
are more than 50 profiles (so the wasted time is not felt). Otherwise, it
just prints a statement that profiles are being built.