Preparing for version 0.13, new --pixelscale in astfits
With this commit, Gnuastro is being prepared for version 0.13.

In the middle, the issue of reading the pixel-scale from any FITS image
came up in another discussion (because some FITS creators don't write the
CDELT keywords!). Since it was very easy to implement, I also added an
option to do this in the FITS program. Ideally those parts should have been
done in a separate commit, but I had already staged some of the changes for
version 0.13 and didn't have the time to re-organize the staged files.