MakeProfiles: corrected uninitialized warning
The newly added 'origquiet' variable in the 'mkprof' function of
MakeProfiles was causing a compilation warning on the possibility of using
an uninitialized value. Practically this will never happen, but the
Compiler has no way of knowing the other parameters.

So to avoid annoying compilation warnings, we are initializing 'origquiet'
in the same time we are defining it.

This issue was reported by Sachin Kumar Singh.