Query: new program to extract data from external databases
Many online databases exist in astronomy that provide command-line access
to their data, enabling users to submit queries and extract a sub-set of
them. Usually people do this on the webpages of the various datasets, but
due to the varying format of each dataset, it is annoying.

Hence with this commit, Gnuastro builds a new program called 'astquery', it
has an internal set of databases and provides a high-level interface to
directly download the desired data on the command-line.