Query: new --width option to automatically select box region
In the last commit, the new 'astquery' program was added, but it could only
select the region around the center with a radius (using '--radius'). With
this commit, it also accepts a new '--width' option, allowing the users to
select rows with a box centered on the given coordinate.

Also, to be more similar to programs like 'asttable' and 'aststatistics',
the short version of the '--column' option is now set to '-c' (previously
it was '-C'), and the short version of '--center' has been changed to '-C'
(was previously '-c').

The correction of option names was suggested by Samane Raji, and Sabastian
is now working with me to get Gnuastro installed in Anaconda.