Library (txt.h): floating point columns, accept sexagesimal values
Until now, it was not easily possible to convert a single sexagesimal
coordiante into degrees on a single command (without making a new file to
add the string metadata). Also, in plain-text tables that had sexagesimal
coordinates, we needed to add the column metadata for the conversion.

With this commit, if the sexagesimal coordinates are written in the
standard format ('_h_m_s' for RA and '_d_m_s' for Dec), they can be present
in a column that is mean to be 32-bit or 64-bit floating point. The
sexagesimal values will be directly converted to a single floating point
number and stored in memory like that.

This couldn't be done in the old sexagesimal format of '_:_:_' because its
ambiguous. So the lower-level unit conversion functions were also changed
to always return the non-ambiguous format, but they can optionally return
the colon-based format with an argument.