Fits: new '--skycoverage' option to report image WCS coverage
Until now it was inconvenient to calculate the the relative coverage of an
image in the sky: where is its center, how big is it (on the sky) and
etc. You had to read the image size, construct a set of 4 coordinates of
the sides of the image, then use 'asttable' and the column arithmetic
operator 'img-to-wcs'.

With this commit, this process is now formalized inside the Fits program
with the '--skycoverage' option which will return two basic ways to
simplify the coverage of the image over the celestial sphere.