Segment: final labels are now thread-safe and reproducible
Until now, the final labeled images of Segment were not thread-safe, hence
when run on multiple threads, the resulting labels were not reproducible:
the same clumps/objects would get different labels in each run! This is
beacause Segment finds the clumps and objects over each detection in
parallel and there is no way to guarantee which thread is run at which
moment, without drastically decreasing the efficiency of threads!

With this commit, the problem is fixed with a simple strategy: after the
final labels are found, all the object labels (or clumps if Segment is run
in '--onlyclumps' mode) are reset based on their position in the image. In
short, we just parse the pixels from the bottom-left to the top-right give
new labels in this order. This ensures the final output labels are the same
in any run.

This bug was reported by Joanna Sakowska.

This fixes bug #59017.