MakeProfiles: --replace option uses the maximum value
Until now, the '--replace' option would always just put values in the same
order that they came over the final canvas. However, in a multi-threaded
scenario, the order of the written profiles on the final canvas changes in
every run. As a result the output with '--replace' wasn't reproducible over
the overlapping pixels.

With this commit, instead of using the final value that is written over the
pixel, we use the maximum value over that pixel. Because the maximum value
doesn't depend on the order, the result is thus thread-safe.

This bug was reported by Samane Raji and Zahra Sharbaf.