Query: new --range option to limit rows of the query
Until now, when constructing the query automatically, query would return
the full range of the requested columns within the given spatial range. But
in many situations, we don't need all the rows, we only need rows with
certain values in some columns.

With this commit there is now a new '--range' option that is designed for
this purpose. It can be called multiple times to set ranges in multiple
columns.

Also, the book, the Query program has been moved into the "Data containers"
chapter, because it fits better into the context of this chapter.