Query: new --overlapwith option to automatically determine region
Until now, if you wanted to query a certain database based on an image, you
would have to run 'astfits' with '--skycoverage', then pass the values. But
this is very inconvenient (especially when you are on the command-line and
want a fast result).

With this commit, the main workhorse of the '--skycoverage' option of
'astfits' has been moved inside the WCS library (and is used by
'astfits'). But because its in the library, it can now be automatically
used by 'astquery' too with the new '--overlapwith' option.