Library (txt): integers are only read in base 10
Until now, if an integer was written in the format of '008' for example to
fill empty character columns (which is done by some databases, like
http://www.astro.gsu.edu/wds), the values would be interpretted in octal
format (the standard C library behavior of 'strtol'). But rarely do
datasets contain octal values, so this caused wrong reading of the values
or crashs (when there were numerical characters equal and larger than 8).

So with this commit, the third argument of 'strtol', which is the base of
the string, is set to '10' so it is always interpretted as base-10. Later,
if it becomes necessary to deal with octal strings (or any other base), we
can define a special option for the base of integers.