Match works with blank coordinates, Table can remove blank rows
Until now, when there were blank values in the given coordiantes, Match
would not find the correct row to match, even if it existed! The reason was
that the GSL library (which we used to sort by index) would not properly
account for NaN values.

With this commit to avoid the problem in GSL, NaN values are replaced with
the largest possible floating point number and the problem is fixed. Some
tests have also been added to avoid NaNs in other columns.

In parallel to this, the problem of removing rows from a table with NaN
values in some columns was a very generic one and commonly necessary. So
the Table program now has a new option called '--noblank', which takes the
names of any column that should not contain a NaN. It will then remove
those rows.

In order to do the steps above, two new functions have been added to
Gnuastro's library: 'gal_blank_flag_remove' and 'gal_blank_remove_rows'.

This fixes bug #59155 (Match not working with blank coordinates).

This bug was reported by Zahra Sharbaf.