Query: functions for querying Gaia moved to gaia.c
So far, the only database that 'astquery' can submit queries to is
Gaia. Because of this, the functions to work on the Gaia catalog were
defined in the core `query.c' file. However, we will soon be adding many
more databases, and its hard to manage them all in one file.

With this commit, to allow easy addition of new databases, the Gaia-specfic
functions have been moved to 'gaia.c' and 'gaia.h'.