Crop: added final check to see if there is overlap
Until now, we would simply use the WCS to see if there is any overlap and
assumed its fine afterwards. But in some scenarios, where there is one or
two pixels offset, while the WCS condition finds an overlap, the actual
pixels fail by one pixel or so and don't find any. This caused an error in
the running of Crop.

With this commit, after attempting the crop, do one extra check to make
sure an actual overlap can be found.