MakeCatalog: new --areahalfsum option
With this new option, it is now possible to see how many pixels constitute
half the total brightness (sum of values) within a labeled region (object
or clump). This can be used to get a zero-th order approximation of the
effective radius for example. But should be used with care on "objects"
(which may have multiple peaks), it is perfect for "clumps" though (which
only have one peak).

This was proposed by Ignacio Trujillo.

This commit completes task #15786.