MakeCatalog: new columns for fractional radius, area, and SB
With the addition of the previously added '--areahalfsum' option, Ignacio
Trujillo suggested to calculate the approximate radius that corresponds to
the area, so the new '--halfradiusobs' option was added, and to help in
sorting the option names with the output of '--help', the original option
was renamed to '--halfsumarea'. Also, by dividing the flux with the region
and its area, we now have a new '--halfsumsb' option to report the surface
brightness.

But to help in identifying the slope of objects, and in particular stars in
an image, two other options were added for the user to optionally ask for
areas and radii at custom fractions of total flux: '--fracsumarea1',
'--fracsumarea2', '--fracradiusobs1' and '--fracradiusobs2'. The custom
fractions can be set with the newly added '--fracsum' option (which can
take two floating point values).

These columns were suggested by Ignacio Trujillo.