MakeCatalog: new option to estimate the observed FWHM
The newly added half-sum and frac-sum radius columns of MakeCatalog depend
on the total flux, which is hard to estimate without deblending. So using
them had limitations.

However, the FWHM is much more robust to deblending (at least when the
deblending level is below half the maximum!), so it is now added following
a similar formalism as the half-sum and frac-sum columns (using the area).