MakeCatalog: new --areaarcsec2 option for area in arcsec squared
Until now, it was annoying to get the pixel scale separately, then multiply
its squared value to the area column to have the area in a more portable
unit of arcsec^2.

With this commit, this can now be done internally and the user can ask for
it with the new '--areaarcsec2'.