MakeCatalog: new columns related to maximum value and half of it
Recently the '--fwhmobs' option was added, which can be a very good measure
of the flux distribution within a labeled region. However, we didn't have
any method to access its reliability.

Therefore with this commit there is a new column called '--maximum', which
returns the exact maximum used by the FWHM measurement. By comparing the
value in '--maximum' with the local standard deviation we can see how
reliable the result is. This is described fully in the book under
'--fwhmobs'.

Also, since the maximum value is much less affected by noise or deblending
issues (compared to the total flux), the area of the pixels above this
value can also shed some light on the profile shape and surface
brightness. So three more columns were added for such measurements:
'--halfmaxarea', '--halfmaxsum' and '--halfmaxsb'.