MakeCatalog: new --surfacebrightness column
The surface brightness of a labeled region (in units of magnitude/arcsec^2)
is sometimes necessary but until now, we needed to do the conversion
manually using the brightness and area columns. So with this commit, a new
'--surfacebrightness' column is added that will do the job internally.

Also, I noticed a mistake in the recently added columns '--halfmaxsb' and
'--halfsumsb' (for the surface brightness at half the maximum, and the
surface brightness at half the total sum). I was mistakenly simply dividing
the magnitude by the area (in arcsec^2)! So with this commit, this issue is
also fixed.

Finally, to clarify the definition of surface brightness (in particular for
the description of the surface brightness related columns), a description
of surface brightness was added just after the definition of magnitude in
the book.

This issue was fixed with the help of Ignacio Trujillo.