Webpage: link to tarball of most recent commit added
Just before merging any commit into Gnuastro's 'master' branch, I do a test
build, and run 'make check' to make sure everything is workingly finely. As
a side-effect (since a tarball is created in the process), I put the most
recent tarball in a fixed URL that people who find bugs or suggest tasks
can get the version with their suggestion implemented.

Until now, but this URL is useful for other people who always want the most
recent version, or when they loose/forget the URL address. So In the "Test
releases" section of the webpage, I added a few sentences and put the URL
there so it is publicly available.

This was suggested by Joanna Sakowska.