Webpage: added URL to most recent version of book with tarball
In the previous commit, I added the URL of the tarball with the most recent
version. But its good to also point viewers to the book/documentation of
that version, and what has been added since the last major release.

This has been done with this commit with two links, one to the PDF book
(also on 'http://akhlaghi.org') and another to the most recent NEWS file.