Book: further clarifications on the surface brightness
Following Raul's correction in the previous commit, I noticed that I had
confused dimension-less units with unitless quantities! We have many units
that are also in logarithmic scale, for example the decibel. For more see
https://en.wikipedia.org/wiki/Logarithmic_scale#Logarithmic_units. This
confusion (on my side) was partly caused by the Wikipedia page for
astronomical magnitude, that said its is "unitless".

On the other hand, Radians and Steradians (which are actually ratios) are
standard SI "units", they are just dimensionless.

So the term "unitless" has been changed to dimensionless and generally, I
started discussing the coverage of an object on the sky in the standard
Steradians, and went onto describe why we use arcsec^2 instead. Also, to
argue why brightness should be divided by area (instead of magnitude), I
used the logarithmic vs. linear scale.