Library (data.h): arrays first loaded in RAM, then as mmap'd files
Until now, an internal array was only allocated in the RAM when its size
was smaller (in bytes) than the value given to the '--minmapsize'
option. But this was annoying/buggy when the system has enough RAM to keep
large files.

With this commit, all Gnuastro programs will first attempt to write the
array in RAM, only when it fails (there is no more RAM left), will they use
a memory-mapped file (which can dramatically slow down the program). This
has been done by setting the default value to '--minmapsize' to an
extremely large value, but adding a check after attempting to allocate the
file. If the allocation fails, a memory-mapped file will be created.

In the process, I also noticed that the arithmetic multi-operand operators
for quantile and sigma-clipping weren't actually freeing one of their
internal arrays, causing some wasted program memory. This has been
corrected.

This issue was found thanks to the Euclid OU-MER team and in particular
on data provided by Martin Kuemmel.