Fits: pixel scales also shown in arcsec/pixel
Until now, the recently added '--pixelscale' option would only print the
pixel scale in units of 'deg/pixel', which is the raw units of the CDELT
keyword in the FITS WCS standrad.

However, when the WCS axis is in units of degrees, it is most common to
express the pixel scale in units of 'arcsec/pixel'. So besides printing the
'deg/pixel', when the units of the axis is in units of degrees, the pixel
scale is also printed in units of 'arcsec/pixel'.

Also, with this commit, the pixel area (on an image, or on each slice of a
3D cube) is printed and when we have a 3D dataset, the pixel volume is also
printed. Similar to above, when the units are in 'deg', values are also
printed in 'arcsec' and when we have meters ('m'), the value is also
printed in Angestroms.