MakeNoise: new --bgisbrightness option to interpret --background
Until now, the value to the '--background' option would always be
interpretted as a magnitude. But the magnitude always needs a zeropoint and
in some scenarios (raw tests), we already have the background value in
brightness. Hence converting the background value to magnitudes was just an
extra annoying step.

With this commit, MakeNoise has the new '--bgisbrightness' option to let
the user inform the program that the value given to '--background' should
be interpreted as brightness, not magnitude.