MakeNoise: error printed with background is less than one
Until now there was no sanity check on the value to the given
background. It may happen that the user mistakenly gives a value less than
1. But since hte background is used in a Poisson distribution (which is a
counting distribution), it must be larger than one.

With this commit, a sanity check has been added to check for this. Also
some cleaning up was done in the code and the default values to the
'--background', '--zeropoint' and '--instrument' options has been removed:
the user should specify the type of the background/noise on every
run. Should it be Poisson? Is it in magnitudes?