Book: Some corrections in the first tutorial
Until now, the printed outputs of some of the programs in the first
tutorial had some small inconsistancies: 1) the dates (which were set to be
those where Sufi would have been alive in) were mistakenly "Apr 6" instead
of "Oct 6". 2) the description of the '--ccol' options in the output of
'asmkprof -P' didn't correspond to what is printed when you run it now.

With this commit, those inconsistancies have been fixed.

Also with this commit, the built 'astquery' program has been added to
'.gitignore' since its not a hand-written source file, but a binary and
built file.