Arithmetic: new interpolation operators based on region borders
Until now, to interpolate saturated stars in a dataset, we needed to use
'interpolate-maxngb'. This operator looks within a given number of
neighboring pixels and is more useful in small noisy regions. But as the
blank regions become larger, it can cause a fragmentation in the centers.

With this commit, two new operators are defined 'interpolate-minofregion'
and 'interpolate-maxofregion'. With these two, all blank regions will be
interpolated with a single value that is either the minimum or maximum
value of the non-blank pixels that are immediately bordering it.