MakeCatalog: positions of min and max value account for many pixels
Until now, in the recently added columns like '--maxvx' and '--maxvy' we
were mistakely reporting the postiion of the first pixel that has the
maximum value. But in scenarios where the centers of saturated stars may be
saturated (and they are replaced by the maximum value), this would cause
significantly different positions (when there is a saturation spike in
particular).

With this commit, the mean position of all the maximum pixels is returned,
not just the first one. This fixes the problem above.

Also, I added to new columns: '--areaminv' and '--areamaxv' to report the
number of pixels that have the minimum/maximum value. This can be useful in
identifying saturated stars in the scenarios above.