MakeCatalog: object ID mapping now applied to corrected clump labels
Until now, When MakeCatalog is run on an objects labeled image that has
non-contiguous labels (for example 23, 25, 800: usually happens when making
crops of labeled images), it applies an internal mapping of the actual
labels/IDs with a contiguous set of IDs. But this scenario (of
non-contiguous IDs) was only tested on generating object catalogs. While
working on a cropped clumps image (that was accompanied with a cropped
label and values image), I noticed that MakeCatalog crashes with a
segmentation fault.

With this commit this issue has been fixed: I noticed that the mapping the
object IDs wasn't being used in the part where we are identifying the
clumps, so they were trying to associate themselves with regions far
outside the allocated number of objects in memory.

This fixes bug #59371.