Book: Added example for Arithmetic's interpolate-medianngb
Until now, the 'interpolate-medianngb' operator of the Arithmetic program
didn't have an example, making it hard to use for someone unfamiliar (it
took me many trials and errors!).

With this commit, an example has been added to this operator for further
clarification. Also a few typos were fixed in some other parts.