MakeCatalog: fraction columns now based on maximum, not sum
Until now, we had some good options to sample the profile at various
fractions of the total sum of the labeled region's pixels. However, the
total sum can easily be biased (for example by blending with other
objects). So its not really too useful.

A more useful measure that we have found in our tests is some measurements
on the fractions of the maximum (like the FWHM). Because the maximum as the
largest per-pixel S/N, it is a very robust measure without much scatter.

With this commit, the old '--fracsum____' columns have been renamed to
'--fracmax____' and two new columns are now also available: '--fracmaxsum1'
and '--fracmaxsum2'.