MakeCatalog: corrected metadata of fractional maximum values
Until now, the metadata of the 'FRAC_MAX_SUM_1' and 'FRAC_MAX_SUM_2'
columns was wrong in both the numerical data type of the column (it was
mistakenly written as an integer), and the column comment didn't contain
the number of fraction.

With this commit, the numeric data type is now 32-bit floating point and
the description is also customized for the number of the fraction.

This issue was reported by Joanna Sakowska.