MakeProfiles: new ability to build any custom radial profile
Until now, MakeProfiles could only build profiles with a pre-defined
parametric function. But in some scenarios users need to build custom
profiles and using the output of the distance profile may not be enough.

With this commit, MakeProfiles now has a new 'custom' profile that comes
with the new '--customtable' and '--customtablehdu' options. For such
profiles the a 3-column table will be necessary that has the interval's
minimum radius, maximum radius and value to use.