Bootstarp: Author list can be generated with no tags
Until now, if the git history didn't contain tags the bootstrapping failed
because of 'git --describe' crashed. Also when 'genauthors' crashed the
bootstrapping would continue and crash much later.

With this commit, an 'if' statement has been defined so when the
'genauthors' crashed, the bootstrapping stops. Also to fix the probem with
git an '--always' was added.

I also recognized that when the bootstrapping is manually aborted, a
temporary file is created so I added to '.gitignore'.