Book: examples for many Arithmetic's operators
Until now, some of the Arithmetic's operators had examples and most of the
more basic ones didn't. However, or a first-time user, those basic ones are
the most familiar and without examples, adapting Arithmetic can be hard.

With this commit, to making usage of the Arithmetic's operator easy for
someone unfamiliar, examples are given for many of Arithmetic's operators
which did not have an example (even the simpler ones).