NEWS file: added item for new Arithmetic examples
In the previous commit, many examples were added to the description
arithmetic operators which can be useful for new users.

With this commit, an entry was added in the NEWS file to highlight this
point there (which also goes into the announcements).

I also done a simple spell-check in the NEWS file and corrected a few
typos. Furthermore, after a run of Debian's Lintian, a few other typos were
found in the printed messages of some of the programs that has been
corrected.