Library and book: memory management by checking available RAM
Until recently, the main way to manage the memory was to set
'--minmapsize', but for that the user needed to know/worry-about the size
of the arrays. So a feature was added to read the available space on RAM
before every allocation. However, in Linux, there is no system call for
this, so we have to read from a file and this made some programs (which
just need to allocate very small arrays) very slow.

With this commit, things have been cleaned up into modular functions (some
of them new) to help in the management. Now, the available memory is only
checked when we array to allocate is larger than 10MB. Also the "Memory
management" section of the book has been updated to be more clear.