Query: when no --output specified, name includes DATASUM of data
Unlike all the other Gnuastro programs, that have an input file, Query
doesn't have any. So its output filename was simply 'query.fits'. But given
that it also contacts an outside server (which may not be available later),
having such a generic name (which can easily be over-written by a new call)
could cause the loss of data and major frustration.

Also, until now, the raw downloaded file was given to the user. However, we
recently discovered that Gaia provides compressed FITS files, so a '.fits'
name could be confusing (even though technically it has no problem within
Gnuastro).

To fix both problems, with this commit, the raw data file is actually
written into a temporary named file (that is unique and will certainly not
exist, thanks to the C library's 'mkstemp' function). That file is then
read into memory and written into an output table. In the process, when no
output name is provided, the 'DATASUM' of the downloaded data (part of the
FITS standard) is calculated and written into the name of the output file.
The extra processing steps will take negligible processing power compared
to contacting a remote server, so they don't add any extra burden.

The fact that we read the table and write it out, allows the produced
tables from any database to have a similar/unified format for the
users. For example, when the output is a FITS table in the zero-th/first
extension, there will be keywords containing the values of all the given
options to Query (so they can "reproduce" the result later: "reproduce"
because something may change on the server, but that is beyond our
control).

This was suggested by Francois Ochsenbein.