Book: added note on Query's --overlapwith and images with bad WCS
Until now, there was no mention of WCS problems in the Query program's
'--overlapwith' option (in particular that the reference point isn't within
the image when there is distortion).

So with this commit, an explanation has been added to warn the users about
such scenarios. Later, we should add some checks for this.