Book: added paragraph on env-based shebang in tutorial on scripts
Until now, in the toturial on shell scripting, we had only mentioned the
'#!/bin/bash' shebang. But it is also common to use '#!/usr/bin/env bash'
to use the first Bash in PATH.

With this commit, to help new users also get familiar with the env-based
shebangs, a paragraph is added that describes the situations it can be
useful. It concludes with an explanation that for simplicity we'll continue
with the direct bash call here.

This was suggested by Carlos Morales Socorro.