CosmicCalculator: fix bug in calling --printparams
Until now, when CosmicCalculator was called with the '--printparams' or
'-P' option it aborted with an error about not setting the redshift. Whilst
the redshift value is not mandatory when the user is asking for checking
the configured parameters.

With this commit, a condition has been added on the check for redshifts at
the start of the program to let the program continue.

This fixes bug #59400.