Library (arithmetic): no extra freeing in sigclip and quantile
Until now, when the Arithemtic program was run with the sigclip-* or
quantile operators it would crash on some systems with a double free
error. The reason was a recently added 'free(pixs)' command that was added
after these parameters because I mistakenly thought that we are not freeing
them. However, the allocated pointers were set inside the 'cont' array and
freed within that.

With this commit, the extra 'free(pixs)' command has been removed and a
comment has been added to remind future readers of the code about the
possible confusion.

This issue was reported by Alberto Madrigal.