NoiseChisel: corrected compiler warning
Until now, the printing (when '--quiet' isn't ginve) of completion of the
growing phase was being reported just after the function that done the
actual growing. While testing the build on an older computer, I saw a
compiler warning about a possibly un-initialized variable.

To keep things clean, the printing of the status is now taken within the
same 'if' statement that does the growing, thus removing this warning, and
also making the code more cleaner.