Crop: new option to write crop in primary (0-th) HDU
Until now, like all programs, Crop would leave the first (0-th) HDU empty
of data (only for keeping metadata) and write the cropped image into the
next HDU. This is suggested by the FITS standard and is good for storing
meta data like configuration parameters. But in some scenarios (especially
when feeding into non-Gnuastro programs that only accept input in the 0-th
extension), this can be annoying.

With this commit the Crop program has a new option called '--primaryimghdu'
(same name of similar option in the Fits program). When this option is
given, the output crop will be written into the primary extension.

This was relatively easy to do for Crop in particular because Crop doesn't
use the default Gnuastro library function to write FITS files, *and* Crop
doesn't have too much configuration and writes its basic metadata in the
same extension as the data (it only has an output of one extension). For
other programs this will be more tricky to implement.