Bootstrapping: images made similar to info or man pages
Until now, we were building the images that go into the webpage and PDF
book at the end of the 'bootstrap' script. But the bootstrap script is
primarily for core programming components of Gnuastro. However, images in
the book are like the man pages or info pages that should be built from the
version controlled source and inserted into the tarball.

With this commit, the call to build the images is now done within
'doc/Makefile.am' step (during the 'make' phase, if they don't exist). But
the images are already packaged into the tarball, so they won't cause a
problem for a user unpacking the tarball.

Also, while cleaning 'boostrap.conf', I noticed that the 'fdl' package is
suggesting to include 'fdl.texi' in the version controlled source. So this
is also done in this commit.

This was suggested by Bruno Haible.