Bootstrap: removed dependency on configure script to build images
In the previous commit, where building images has been moved to
'doc/Makefile.am', I had set the building of images to be dependent on the
'configure' script. But that is not the case here (unlike
'authors.texi'). As a result, the tarball wouldn't work.

With this commit, the images don't depend on anything. If they exist, they
will not be rebuilt, and if they don't, they will.

This bug was reported by Samane raji