NoiseChisel & Statistics: improved outlier rejection in Sky
Until now, once the detection was done, the tiles that weren't
significantly covered by detections (based on '--minskyfrac') would be
directly used for the interpolation phase: there was no outlier rejection
beyond the median-based interpolation. Also, when removing outliers during
the finding of the quantile threshold, we were using global measures to
find and reject outliers. So it wouldn't opeate too well when there were
strong gradients for example.

With this commit, to find outliers in the quantile threshold, we now
calculate the distance between the second-smallest and second-largest
values in the '--interpnumngb' tiles around it. This allows for a more
local rejection of some tiles. This is done in the newly added
'gal_tileinternal_no_outlier_local' funciton. We then look at the
distribution of those values to find outliers. Also, outlier rejection is
now also implemented on the final Sky image.

In the process, some other issues have been fixed:

 - While looking at previous code to remember the low-level implementation
   details, I noticed that the lines are longer than 75 characters, making
   the functions hard to read.

 - The internal 'INTERPOLATE_FLAGS_CHECKED' macro in 'lib/interpolate.c'
   was renamed to 'INTERPOLATE_FLAGS_NGB_CHECKED' to be more clear. When
   reviewing the code, initially I mistakenly thought it was the fact that
   a pixel is checked or not. But it was actually a check to see if the
   pixel has been checked during the search to find the nearest neighbor.

 - In the 'gal_statistics_quantile_function_index' function, I noticed that
   we weren't accounting for cases where the lowest two elements have the
   same value (to see if the array is sorted in increasing or decreasing
   order). As a result, some results were being mistakenly returned as
   'inf'. With this commit, we are now using the relevant flag that
   explicitly keeps this information.

 - Added sanity checks on the 'value' argument in
   'gal_statistics_quantile_function' and its index finding function.